+++ src/renderer/Metadata.js
@@ -19,17 +19,23 @@ export default class Metadata {
 
         Object.values(properties).map(property => {
             property.categories = property.categories || [];
-            property.categories.map(category => this.categorizeString(category.name));
+            property.categories.map(category => this.categorizeString(property, category.name, true));
         });
 
         this.propertyKeys = Object.keys(this.properties);
     }
-    categorizeString (category) {
+    categorizeString (propertyName, category, init = false) {
         if (category === undefined) {
             category = null;
         }
         if (this.categoryToID.has(category)) {
             return this.categoryToID.get(category);
+        }
+        if (!init) {
+            this.properties[propertyName].categories.push({
+                name: category,
+                frequency: Number.NaN
+            });
         }
         this.categoryToID.set(category, this.numCategories);
         this.IDToCategory.set(this.numCategories, category);