+++ src/renderer/viz/expressions/aggregation/cluster/ClusterSum.js
@@ -1,5 +1,5 @@
 import ClusterAggregation from './ClusterAggregation';
-
+import { checkMaxArguments } from '../../utils';
 /**
  * Aggregate using the sum. This operation disables the access to the property
  * except within other cluster aggregate functions.
@@ -27,6 +27,7 @@ import ClusterAggregation from './ClusterAggregation';
  */
 export default class ClusterSum extends ClusterAggregation {
     constructor (property) {
+        checkMaxArguments(arguments, 1, 'clusterSum');
         super({ property, expressionName: 'clusterSum', aggName: 'sum', aggType: 'number' });
     }
 }