+++ src/renderer/viz/expressions/classification/ViewportEqIntervals.js
@@ -1,6 +1,6 @@
 import Classifier from './Classifier';
 import { viewportMax, viewportMin } from '../../expressions';
-import { checkNumber, checkType } from '../utils';
+import { checkNumber, checkType, checkMaxArguments } from '../utils';
 
 /**
  * Classify `input` by using the equal intervals method with `n` buckets.
@@ -29,6 +29,7 @@ import { checkNumber, checkType } from '../utils';
  */
 export default class ViewportEqIntervals extends Classifier {
     constructor (input, buckets) {
+        checkMaxArguments(arguments, 2, 'viewportEqIntervals');
         checkNumber('viewportEqIntervals', 'buckets', 1, buckets);
 
         const children = { input, _min: viewportMin(input), _max: viewportMax(input) };