+++ test/integration/user/test/interactivity.test.js
@@ -197,90 +197,104 @@ describe('Interactivity', () => {
                 });
             });
         });
+    });
 
-        describe('when the user move the mouse on the map', () => {
-            describe('and the mouse enters in a feature', () => {
-                it('should fire a featureHover event with a features list containing the entered feature', done => {
-                    interactivity.on('featureHover', event => {
-                        expect(event.features[0].id).toEqual(-0);
-                        expect(event.features[0].layerId).toEqual('layer1');
-                        done();
-                    });
-                    onLoaded(() => {
-                        // Move mouse inside a feature 1
-                        util.simulateMove({ lng: 5, lat: 5 });
-                    });
+    describe('when the user moves the mouse on the map', () => {
+        describe('and the mouse enters in a feature', () => {
+            it('should fire a featureHover event with a features list containing the entered feature', done => {
+                interactivity.on('featureHover', event => {
+                    expect(event.features.length).toBe(1);
+                    expect(event.features[0].id).toEqual(-0);
+                    expect(event.features[0].layerId).toEqual('layer1');
+                    done();
                 });
-
-                it('should fire a featureEnter event with a features list containing the entered feature', done => {
-                    interactivity.on('featureEnter', event => {
-                        expect(event.features[0].id).toEqual(-0);
-                        expect(event.features[0].layerId).toEqual('layer1');
-                        done();
-                    });
-                    onLoaded(() => {
-                        // Move mouse inside a feature 1
-                        util.simulateMove({ lng: 5, lat: 5 });
-                    });
-                });
-
-                it('should not fire a featureEnter event when the mouse is moved inside the same feature', done => {
-                    const featureEnterSpy = jasmine.createSpy('featureEnterSpy');
-                    onLoaded(() => {
-                        interactivity.on('featureEnter', featureEnterSpy);
-                        // Move mouse inside a feature 1
-                        util.simulateMove({ lng: 5, lat: 5 });
-                        // Move mouse inside the same feature 1
-                        util.simulateMove({ lng: 5, lat: 15 });
-                        setTimeout(() => {
-                            expect(featureEnterSpy).toHaveBeenCalledTimes(1);
-                            done();
-                        }, 0);
-                    });
+                onLoaded(() => {
+                    // Move mouse inside a feature 1
+                    util.simulateMove({ lng: 5, lat: 5 });
                 });
             });
 
-            describe('and the mouse leaves a feature', () => {
-                it('should fire a featureHover event with an empty features list', done => {
-                    onLoaded(() => {
-                        // Move mouse inside a feature 1
-                        util.simulateMove({ lng: 5, lat: 5 });
-                        interactivity.on('featureHover', event => {
-                            expect(event.features.length).toEqual(0);
-                            done();
-                        });
-                        // Move mouse outside any feature (over a feature 1 hole)
-                        util.simulateMove({ lng: 15, lat: 15 });
-                    });
+            it('should fire a featureEnter event with a features list containing the entered feature', done => {
+                interactivity.on('featureEnter', event => {
+                    expect(event.features.length).toBe(1);
+                    expect(event.features[0].id).toEqual(-0);
+                    expect(event.features[0].layerId).toEqual('layer1');
+                    done();
                 });
-
-                it('should fire a featureLeave event with a features list containing the previously entered feature', done => {
-                    onLoaded(() => {
-                        interactivity.on('featureLeave', event => {
-                            expect(event.features[0].id).toEqual(-0);
-                            expect(event.features[0].layerId).toEqual('layer1');
-                            done();
-                        });
-                        // Move mouse inside a feature 1
-                        util.simulateMove({ lng: 5, lat: 5 });
-                        // Move mouse outside any feature
-                        util.simulateMove({ lng: -5, lat: -5 });
-                    });
+                onLoaded(() => {
+                    // Move mouse inside a feature 1
+                    util.simulateMove({ lng: 5, lat: 5 });
                 });
+            });
 
-                it('should not fire a featureLeave event when the mouse is moved outside any feature', done => {
-                    layer2.on('loaded', () => {
-                        const featureLeaveSpy = jasmine.createSpy('featureLeaveSpy');
-                        // Move mouse outside any feature
-                        util.simulateMove({ lng: -5, lat: -5 });
-                        interactivity.on('featureLeave', featureLeaveSpy);
-                        interactivity.on('featureHover', () => {
-                            expect(featureLeaveSpy).not.toHaveBeenCalled();
-                            done();
-                        });
-                        // Move mouse outside any feature
-                        util.simulateMove({ lng: -10, lat: -10 });
+            it('should not fire a featureEnter event when the mouse is moved inside the same feature', done => {
+                const featureEnterSpy = jasmine.createSpy('featureEnterSpy');
+                onLoaded(() => {
+                    interactivity.on('featureEnter', featureEnterSpy);
+                    // Move mouse inside a feature 1
+                    util.simulateMove({ lng: 5, lat: 5 });
+                    // Move mouse inside the same feature 1
+                    util.simulateMove({ lng: 5, lat: 15 });
+                    setTimeout(() => {
+                        expect(featureEnterSpy).toHaveBeenCalledTimes(1);
+                        done();
+                    }, 0);
+                });
+            });
+        });
+
+        describe('and the mouse leaves a feature', () => {
+            it('should fire a featureLeave event with a features list containing the previously entered feature', done => {
+                onLoaded(() => {
+                    // Move mouse inside a feature 1
+                    util.simulateMove({ lng: 5, lat: 5 });
+                    interactivity.on('featureLeave', event => {
+                        expect(event.features.length).toBe(1);
+                        expect(event.features[0].id).toEqual(-0);
+                        expect(event.features[0].layerId).toEqual('layer1');
+                        done();
                     });
+                    // Move mouse outside any feature
+                    util.simulateMove({ lng: -5, lat: -5 });
+                });
+            });
+
+            it('should not fire a featureLeave event when the mouse is moved outside any feature', done => {
+                layer2.on('loaded', () => {
+                    const featureLeaveSpy = jasmine.createSpy('featureLeaveSpy');
+                    // Move mouse outside any feature
+                    util.simulateMove({ lng: -5, lat: -5 });
+                    interactivity.on('featureLeave', featureLeaveSpy);
+                    interactivity.on('featureHover', () => {
+                        expect(featureLeaveSpy).not.toHaveBeenCalled();
+                        done();
+                    });
+                    // Move mouse outside any feature (over a feature 1 hole)
+                    util.simulateMove({ lng: 15, lat: 15 });
+                });
+            });
+        });
+    });
+
+    describe('when the layer changes', () => {
+        describe('and appears a feature below the mouse', () => {
+            it('should fire a featureHover event with the feature 1', done => {
+                onLoaded(() => {
+                    // Hide feature
+                    viz1.filter = 0;
+                    // Setup initial mouse position
+                    util.simulateMove({ lng: 5, lat: 5 });
+
+                    // Register event after move to be called by layer `updated`
+                    interactivity.on('featureHover', event => {
+                        expect(event.features.length).toBe(1);
+                        expect(event.features[0].id).toEqual(-0);
+                        expect(event.features[0].layerId).toEqual('layer1');
+                        done();
+                    });
+
+                    // Show feature
+                    viz1.filter = 1;
                 });
             });
         });