+++ test/unit/renderer/viz/expressions/AlphaNormalize.test.js
@@ -1,0 +1,28 @@
+import { validateStaticType, validateMaxArgumentsError, validateCompileTypeError } from './utils';
+import { alphaNormalize, rgb, property } from '../../../../../src/renderer/viz/expressions';
+
+describe('src/renderer/viz/expressions/AlphaNormalize', () => {
+    describe('type', () => {
+        validateStaticType('alphaNormalize', ['color', 'number-property'], 'color');
+    });
+    describe('error control', () => {
+        validateMaxArgumentsError('alphaNormalize', ['color', 'number', 'number']);
+        validateCompileTypeError('alphaNormalize', []);
+        validateCompileTypeError('alphaNormalize', ['color']);
+        validateCompileTypeError('alphaNormalize', ['color', 'category']);
+    });
+    describe('eval', () => {
+        it('should return a normalized color', () => {
+            const expr = alphaNormalize(rgb(255, 128, 0), property('price'));
+            expr._bindMetadata({
+                properties: {
+                    price: {
+                        type: 'number',
+                        max: 10
+                    }
+                }
+            });
+            expect(expr.eval({price: 5})).toEqual({r: 255, g: 128, b: 0, a: 0.5});
+        });
+    });
+});